+++ Solution3/Alignments/untitled/src/Gotoh.java
@@ -1,9 +1,9 @@
 public abstract class Gotoh extends AlignmentAlgorithm {
-    protected int[][] M;
-    protected int[][] I;
-    protected int[][] D;
-    protected int gapopen;
-    protected int gapextend;
+    protected double[][] M;
+    protected double[][] I;
+    protected double[][] D;
+    protected double gapopen;
+    protected double gapextend;
     protected int seqLen;
     protected int seqLen2;
     public Gotoh(Sequence seq1, Sequence seq2, Matrix matrix, GapPenalty gapPenalty) {
@@ -12,9 +12,9 @@ public abstract class Gotoh extends AlignmentAlgorithm {
         this.seqLen = seq1.getLength();
         this.seqLen2 = seq2.getLength();
 
-        M = new int[seqLen+1][seqLen2+1];
-        I = new int[seqLen+1][seqLen2+1];
-        D = new int[seqLen+1][seqLen2+1];
+        M = new double[seqLen+1][seqLen2+1];
+        I = new double[seqLen+1][seqLen2+1];
+        D = new double[seqLen+1][seqLen2+1];
 
         gapopen = gapPenalty.getOpen();
         gapextend = gapPenalty.getExtend();
@@ -30,7 +30,7 @@ public abstract class Gotoh extends AlignmentAlgorithm {
 
         for (int i = 1; i <= seqLen; i++) {
             for (int j = 1; j <= seqLen2; j++) {
-                int score = scoringMatrix.score(seq1.charAt(i-1), seq2.charAt(j-1));
+                double score = scoringMatrix.score(seq1.charAt(i-1), seq2.charAt(j-1));
 
                 I[i][j] = adjust(Math.max(
                         I[i-1][j] + gapextend,
@@ -54,7 +54,7 @@ public abstract class Gotoh extends AlignmentAlgorithm {
         }
     }
 
-    protected int adjust(int value){
+    protected double adjust(double value){
         return value;
     }
